backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-socketio
python-multipart